back/requirements.txt
fastapi==0.110.*           
uvicorn[standard]==0.29.*  
sqlalchemy==2.0.*          
pydantic==2.*               
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.*  
python-multipart==0.0.9   
python-dotenv==1.*         
aiofiles==23.*            
tzdata==2024.* 
pytz
pydantic[email]
pip install slowapi
